refactor: PieceResponse dto 패키지로 분리
