+++ src/main/java/chess/view/OutputView.java
@@ -1,6 +1,6 @@
 package chess.view;
 
-import chess.domain.pieces.piece.PieceResponse;
+import chess.dto.PieceResponse;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.IntStream;
