+++ src/test/java/chess/domain/board/BoardFactoryTest.java
@@ -2,7 +2,7 @@ package chess.domain.board;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.pieces.piece.PieceResponse;
+import chess.dto.PieceResponse;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
